+++ custom_components/prix_carburant/sensor.py
@@ -12,7 +12,7 @@ from homeassistant.components.sensor import (
     SensorEntity,
 )
 from homeassistant.config_entries import SOURCE_IMPORT, ConfigEntry
-from homeassistant.const import ATTR_NAME, CURRENCY_EURO
+from homeassistant.const import ATTR_LATITUDE, ATTR_LONGITUDE, ATTR_NAME, CURRENCY_EURO
 from homeassistant.core import HomeAssistant
 from homeassistant.exceptions import ConfigEntryNotReady
 import homeassistant.helpers.config_validation as cv
@@ -27,10 +27,12 @@ from .const import (
     ATTR_FUELS,
     ATTR_POSTAL_CODE,
     ATTR_PRICE,
-    CARBURANTS,
+    ATTR_UPDATED_DATE,
+    CONF_FUELS,
     CONF_MAX_KM,
     CONF_STATIONS,
     DOMAIN,
+    FUELS,
 )
 from .tools import PrixCarburantTool
 
@@ -70,6 +72,10 @@ async def async_setup_entry(
     config = entry.data
     options = entry.options
     max_distance = options.get(CONF_MAX_KM, config.get(CONF_MAX_KM))
+    enabled_fuels = {}
+    for fuel in FUELS:
+        fuel_key = f"{CONF_FUELS}_{fuel}"
+        enabled_fuels[fuel] = options.get(fuel_key, config.get(fuel_key, True))
 
     tool = await hass.async_add_executor_job(PrixCarburantTool)
 
@@ -107,11 +113,14 @@ async def async_setup_entry(
     _LOGGER.info("%s stations found", str(len(user_stations_ids)))
     entities = []
     for station_id in user_stations_ids:
-        for carburant in CARBURANTS:
-            if carburant in tool.stations[station_id][ATTR_FUELS]:
+        for fuel in FUELS:
+            if (
+                fuel in tool.stations[station_id][ATTR_FUELS]
+                and enabled_fuels[fuel] is True
+            ):
                 entities.append(
                     PrixCarburant(
-                        station_id, tool.stations[station_id], carburant, coordinator
+                        station_id, tool.stations[station_id], fuel, coordinator
                     )
                 )
 
@@ -121,24 +130,24 @@ async def async_setup_entry(
 class PrixCarburant(SensorEntity):
     """Representation of a Sensor."""
 
-    def __init__(self, station_id, station_info, carburant, coordinator):
+    def __init__(self, station_id, station_info, fuel, coordinator):
         """Initialize the sensor."""
         self.station_id = station_id
         self.station_info = station_info
-        self.carburant = carburant
+        self.fuel = fuel
         self.coordinator = coordinator
 
         self._last_update = None
 
         self._attr_icon = "mdi:gas-station"
         self._attr_device_class = SensorDeviceClass.MONETARY
-        self._attr_unique_id = "_".join([DOMAIN, self.station_id, self.carburant])
+        self._attr_unique_id = "_".join([DOMAIN, self.station_id, self.fuel])
         self._attr_native_unit_of_measurement = CURRENCY_EURO
         if self.station_info[ATTR_NAME] != "undefined":
             station_name = f"Station {self.station_info[ATTR_NAME]}"
         else:
             station_name = f"Station {self.station_id}"
-        self._attr_name = f"{station_name} {self.carburant}"
+        self._attr_name = f"{station_name} {self.fuel}"
         self._attr_device_info = DeviceInfo(
             identifiers={(DOMAIN, self.station_id)},
             manufacturer="Station",
@@ -151,12 +160,15 @@ class PrixCarburant(SensorEntity):
             ATTR_ADDRESS: self.station_info[ATTR_ADDRESS],
             ATTR_POSTAL_CODE: self.station_info[ATTR_POSTAL_CODE],
             ATTR_CITY: self.station_info[ATTR_CITY],
+            ATTR_LATITUDE: self.station_info[ATTR_LATITUDE],
+            ATTR_LONGITUDE: self.station_info[ATTR_LONGITUDE],
+            ATTR_UPDATED_DATE: self.station_info[ATTR_UPDATED_DATE],
         }
 
     @property
     def native_value(self):
         """Return the current price."""
-        if self.carburant in self.coordinator.data[self.station_id][ATTR_FUELS]:
-            return self.coordinator.data[self.station_id][ATTR_FUELS][self.carburant][
+        if self.fuel in self.coordinator.data[self.station_id][ATTR_FUELS]:
+            return self.coordinator.data[self.station_id][ATTR_FUELS][self.fuel][
                 ATTR_PRICE
             ]